% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.R
\name{combine_grading_results}
\alias{combine_grading_results}
\title{Combine Grading Results}
\usage{
combine_grading_results(results)
}
\arguments{
\item{results}{List of grading results from individual responses}
}
\value{
Tibble with properly structured results
}
\description{
Properly combine list of grading results into a tibble
}
\keyword{internal}
