% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_optimal_cores}
\alias{get_optimal_cores}
\title{Get Optimal Number of Cores for Parallel Processing}
\usage{
get_optimal_cores(n_tasks, max_cores = NULL)
}
\arguments{
\item{n_tasks}{Number of tasks to process}

\item{max_cores}{Maximum cores to use (optional)}
}
\value{
Recommended number of cores
}
\description{
Determine the optimal number of cores based on system and data size
}
