% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_prompting.R
\name{read_file_all}
\alias{read_file_all}
\title{Read File Contents}
\usage{
read_file_all(p)
}
\arguments{
\item{p}{File path}
}
\value{
File contents as single string
}
\description{
Read File Contents
}
\keyword{internal}
