% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_load_rubric.R
\name{read_md_frontmatter}
\alias{read_md_frontmatter}
\title{Read Markdown Frontmatter}
\usage{
read_md_frontmatter(path)
}
\arguments{
\item{path}{Path to markdown file with YAML frontmatter}
}
\value{
Parsed YAML data
}
\description{
Read Markdown Frontmatter
}
\keyword{internal}
