% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{should_use_parallel}
\alias{should_use_parallel}
\title{Check if Parallel Processing is Beneficial}
\usage{
should_use_parallel(n_tasks, min_tasks_for_parallel = 4)
}
\arguments{
\item{n_tasks}{Number of tasks to process}

\item{min_tasks_for_parallel}{Minimum tasks to justify parallel processing}
}
\value{
TRUE if parallel processing is recommended
}
\description{
Determine if parallel processing would be beneficial for the given task
}
