% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_response_data}
\alias{validate_response_data}
\title{Validate Response Data}
\usage{
validate_response_data(responses)
}
\arguments{
\item{responses}{Data frame with student responses}
}
\value{
TRUE if valid, otherwise throws error
}
\description{
Check that response data has required columns
}
