% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{view_student_results}
\alias{view_student_results}
\title{View Grading Results Details}
\usage{
view_student_results(results, student_index = 1)
}
\arguments{
\item{results}{Grading results from grade_responses()}

\item{student_index}{Row number of student (default: 1)}
}
\value{
List with detailed grading information
}
\description{
Show detailed results for a specific student
}
